Add More Functionality to BlueTooth Class
Added code to handle errors that could be
returned after calling D-Bus methods.

Added code to store away exported interfaces
that implement the org.bluez.AgentManager1 and
org.bluez.ProfileManager1 D-Bus interfaces.

Added code to register an agent with the bluez
agent manager.  This will allow us to handle
pairing requests in case the device the user
wants to connect to hasn't been paired to
this machine.

Added code to register a profile with the
bluez profile manager.  This profile specifies
that this program will handle connection of
a bluetooth serial port device, and provide
a mechanism for bluetoothd to hand off a
socket file descriptor to this program once
a connection has been established.

Added a button to the homeview to connect to the
OBD device.  No action has been implemented, yet.

Added CombBox class derived from Gtk::ComboBox that
will create a combo box with a list of any number
of columns of any type.  This is used by the combo
box that is used to select the bluetooth device the
user wants to connect to.