Add More Functionality to BlueTooth Class
Made BlueTooth class a singleton.  This was necessary because,
according to the Glib documentation, Gio::DBus::InterfaceVTable
objects must either be created globally or in main(), so that it
doesn't cause problems with memory leaks if the objects' lifetimes
are cut short; however, these objects only make sense in the
context of the BlueTooth class because it is implementing D-Bus
interfaces.  By making the BlueTooth class a singleton, the
BlueTooth object exists as a global object as does its private
objects, so I can make the Gio::DBus::InterfaceVTable objects
private to the class.  Since it doesn't really make sense to
instantiate more than one BlueTooth object, the implementation
as a singleton makes sense.

Added code to create a "profile" that will be registered with
the bluez profile manager.  This profile will handle connection
requests made for any bluetooth serial device (uuid = 00001101-
0000-1000-8000-00805f9b34fb).  Bluez will handle any connection
request for such a device, create a socket that can be used with
the device and pass that socket to the profile by invoking the
"NewConnection" method for the profile.

Added code to create an "agent" that will be registered with the
bluez agent manager.  The agent may be invoked to handle pairing
request in case the device we are trying to connect to has not
yet been paired with the machine.  This agent will handle requests
for data needed for pairing, such as a request for a pin code for
the paired device, or a request for authorization to pair the
device with this machine.