Fix Potential Dangling Pointers
Insure that code outside classes cannot end up with a
stale pointer in the event that an object is destroyed
by disconnecting any signals that have a pointer to
the object when the object is destroyed or by insuring
that the external code has the shared pointer to the
object.

The sigc library has a base class called sigc::trackable
that will automatically disconnect any signals created
with a call to sigc::mem_fun.  For this reason, I've
changed many of the areas where I converted sigc::mem_fun
calls to lambdas back to calls to sigc::mem_fun.